Create modified-homepage-index
Here's the mainpage so far, please let me know of any ideas or items that can be added to the page. I am a bit stuck on what new things can be added. Please remember that I have not touched js yet, all js is that of Jo. I will upload the styles.css right after this.